docker-deploy/web-app/requirements.txt
Django>3.0
psycopg2
postgres
django-tempus-dominus==5.1.2.13